fix: update time truncation in database session service (#448)
When truncating to microseconds (to match db max allowed precision) we
need to keep the location of the event. Right now it's overwritten here.

That's the fix for tests mostly, because in real world scenario the
event is created and overwritten here by the same library code, thus it
will be the same location anyway.